Drop legacy integration tests
