--- testdata/envoy-with-ns/filelist.yaml
@@ -1,18 +1,0 @@
-- "."
-- ".helmignore"
-- "Chart.yaml"
-- "README.md"
-- "files"
-- "files/templates"
-- "files/templates/configmap.yaml"
-- "files/templates/deployment.yaml"
-- "files/templates/poddisruptionbudget.yaml"
-- "files/templates/service.yaml"
-- "files/templates/xds.configmap.yaml"
-- "templates"
-- "templates/configmap.yaml"
-- "templates/deployment.yaml"
-- "templates/poddisruptionbudget.yaml"
-- "templates/service.yaml"
-- "templates/xds.configmap.yaml"
-- "values.yaml"
