--- testdata/envoy-with-ns/output/Chart.yaml
@@ -1,18 +1,0 @@
-apiVersion: v1
-appVersion: 1.11.2
-description: Envoy is an open source edge and service proxy, designed for cloud-native
-  applications.
-home: https://www.envoyproxy.io/
-icon: https://avatars0.githubusercontent.com/u/30125649
-keywords:
-- envoy
-- proxy
-maintainers:
-- email: josdotso@cisco.com
-  name: josdotso
-- email: ykuoka@gmail.com
-  name: mumoshu
-name: envoy
-sources:
-- https://github.com/envoyproxy/envoy
-version: 1.9.1
