--- testdata/envoy-with-ns/output/files/templates/configmap.yaml
@@ -1,64 +1,0 @@
-# Source: envoy/templates/configmap.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: testrelease-envoy
-  labels:
-    app: envoy
-    chart: envoy-1.9.1
-    release: testrelease
-    heritage: Helm
-  namespace: foo
-data:
-  envoy.yaml: |-
-    ## refs:
-    ## - https://www.envoyproxy.io/docs/envoy/latest/start/start#quick-start-to-run-simple-example
-    ## - https://raw.githubusercontent.com/envoyproxy/envoy/master/configs/google_com_proxy.v2.yaml
-    admin:
-      access_log_path: /dev/stdout
-      address:
-        socket_address:
-          address: 0.0.0.0
-          port_value: 9901
-
-    static_resources:
-      listeners:
-      - name: listener_0
-        address:
-          socket_address:
-            address: 0.0.0.0
-            port_value: 10000
-        filter_chains:
-        - filters:
-          - name: envoy.http_connection_manager
-            config:
-              access_log:
-              - name: envoy.file_access_log
-                config:
-                  path: /dev/stdout
-              stat_prefix: ingress_http
-              route_config:
-                name: local_route
-                virtual_hosts:
-                - name: local_service
-                  domains: ["*"]
-                  routes:
-                  - match:
-                      prefix: "/"
-                    route:
-                      host_rewrite: www.google.com
-                      cluster: service_google
-              http_filters:
-              - name: envoy.router
-      clusters:
-      - name: service_google
-        connect_timeout: 0.25s
-        type: LOGICAL_DNS
-        dns_lookup_family: V4_ONLY
-        lb_policy: ROUND_ROBIN
-        hosts:
-          - socket_address:
-              address: google.com
-              port_value: 443
-        tls_context:
-          sni: www.google.com
