--- testdata/envoy-with-ns/output/files/templates/xds.configmap.yaml
@@ -1,12 +1,0 @@
-# Source: envoy/templates/xds.configmap.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: testrelease-envoy-xds
-  labels:
-    app: envoy
-    chart: envoy-1.9.1
-    release: testrelease
-    heritage: Helm
-  namespace: foo
-data:
