--- testdata/envoy-with-ns/output/templates/configmap.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/configmap.yaml" }}
