--- testdata/envoy-with-ns/output/templates/deployment.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/deployment.yaml" }}
