--- testdata/envoy-with-ns/output/templates/poddisruptionbudget.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/poddisruptionbudget.yaml" }}
