--- testdata/envoy-with-ns/output/templates/service.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/service.yaml" }}
