--- testdata/envoy-with-ns/output/templates/xds.configmap.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/xds.configmap.yaml" }}
