--- testdata/prometheus-operator-adhoc-dep-with-strategicpatch/filelist.yaml
@@ -1,11 +1,0 @@
-- "."
-- ".helmignore"
-- "CONTRIBUTING.md"
-- "Chart.yaml"
-- "README.md"
-- "files"
-- "files/templates"
-- "files/templates/patched.yaml"
-- "templates"
-- "templates/patched.yaml"
-- "values.yaml"
