--- testdata/prometheus-operator-adhoc-dep-with-strategicpatch/output/templates/patched.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/patched.yaml" }}
