--- testdata/prometheus-operator-adhoc-dep-with-strategicpatch/strategicpatch.yaml
@@ -1,8 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: testrelease-my
-  namespace: default
-spec:
-  strategy:
-    type: RollingUpdate
