--- testdata/prometheus-operator-adhoc-dep-with-strategicpatch/values.yaml
@@ -1,3 +1,0 @@
-my:
-  mysqlRootPassword: uJod1LatG
-  mysqlPassword: uJod1LatG
