--- testdata/prometheus-operator-adhoc-dep/output/Chart.yaml
@@ -1,22 +1,0 @@
-apiVersion: v1
-appVersion: 0.38.1
-description: Provides easy monitoring definitions for Kubernetes services, and deployment
-  and management of Prometheus instances.
-engine: gotpl
-home: https://github.com/coreos/prometheus-operator
-icon: https://raw.githubusercontent.com/prometheus/prometheus.github.io/master/assets/prometheus_logo-cb55bb5c346.png
-keywords:
-- operator
-- prometheus
-maintainers:
-- name: vsliouniaev
-- name: bismarck
-- email: gianrubio@gmail.com
-  name: gianrubio
-name: prometheus-operator
-sources:
-- https://github.com/coreos/kube-prometheus
-- https://github.com/coreos/prometheus-operator
-- https://coreos.com/operators/prometheus
-tillerVersion: '>=2.12.0'
-version: 9.2.2
