--- testdata/prometheus-operator-adhoc-dep/output/files/charts/my/templates/deployment.yaml
@@ -1,105 +1,0 @@
----
-# Source: prometheus-operator/charts/my/templates/deployment.yaml
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: testrelease-my
-  namespace: default
-  labels:
-    app: testrelease-my
-    chart: "my-1.6.6"
-    release: "testrelease"
-    heritage: "Helm"
-
-spec:
-  strategy:
-    type: Recreate
-  selector:
-    matchLabels:
-      app: testrelease-my
-      release: testrelease
-  template:
-    metadata:
-      labels:
-        app: testrelease-my
-        release: testrelease
-    spec:
-      serviceAccountName: default
-      initContainers:
-      - name: "remove-lost-found"
-        image: "busybox:1.32"
-        imagePullPolicy: "IfNotPresent"
-        resources:
-          requests:
-            cpu: 10m
-            memory: 10Mi
-        command:  ["rm", "-fr", "/var/lib/mysql/lost+found"]
-        volumeMounts:
-        - name: data
-          mountPath: /var/lib/mysql
-      # - name: do-something
-      #   image: busybox
-      #   command: ['do', 'something']
-      
-      containers:
-      - name: testrelease-my
-        image: "mysql:5.7.30"
-        imagePullPolicy: "IfNotPresent"
-        resources:
-          requests:
-            cpu: 100m
-            memory: 256Mi
-        env:
-        - name: MYSQL_ROOT_PASSWORD
-          valueFrom:
-            secretKeyRef:
-              name: testrelease-my
-              key: mysql-root-password
-        - name: MYSQL_PASSWORD
-          valueFrom:
-            secretKeyRef:
-              name: testrelease-my
-              key: mysql-password
-              optional: true
-        - name: MYSQL_USER
-          value: ""
-        - name: MYSQL_DATABASE
-          value: ""
-        ports:
-        - name: mysql
-          containerPort: 3306
-        livenessProbe:
-          exec:
-            command:
-            - sh
-            - -c
-            - "mysqladmin ping -u root -p${MYSQL_ROOT_PASSWORD}"
-          initialDelaySeconds: 30
-          periodSeconds: 10
-          timeoutSeconds: 5
-          successThreshold: 1
-          failureThreshold: 3
-        readinessProbe:
-          exec:
-            command:
-            - sh
-            - -c
-            - "mysqladmin ping -u root -p${MYSQL_ROOT_PASSWORD}"
-          initialDelaySeconds: 5
-          periodSeconds: 10
-          timeoutSeconds: 1
-          successThreshold: 1
-          failureThreshold: 3
-        volumeMounts:
-        - name: data
-          mountPath: /var/lib/mysql
-        # - name: extras
-        #   mountPath: /usr/share/extras
-        #   readOnly: true
-        
-      volumes:
-      - name: data
-        persistentVolumeClaim:
-          claimName: testrelease-my
-      # - name: extras
-      #   emptyDir: {}
