--- testdata/prometheus-operator-adhoc-dep/output/templates/charts/my/templates/deployment.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/charts/my/templates/deployment.yaml" }}
