--- testdata/raw-manifests/filelist.yaml
@@ -1,11 +1,0 @@
-- "."
-- "Chart.yaml"
-- "files"
-- "files/templates"
-- "files/templates/configmap.yaml"
-- "files/templates/foo"
-- "files/templates/foo/configmap.2.yaml"
-- "templates"
-- "templates/configmap.yaml"
-- "templates/foo"
-- "templates/foo/configmap.2.yaml"
