--- testdata/raw-manifests/input/configmap.yaml
@@ -1,21 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig
-data:
-  foo: bar
-  bar: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
----
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig2
-data:
-  foo: bar
-  bar: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
