--- testdata/raw-manifests/input/foo/configmap.2.yaml
@@ -1,10 +1,0 @@
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig
-data:
-  foo: baz
-  baz: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
