--- testdata/raw-manifests/output/files/templates/configmap.yaml
@@ -1,25 +1,0 @@
-# Source: input/templates/configmap.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig
-  namespace: foo
-data:
-  foo: bar
-  bar: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
----
-# Source: input/templates/configmap.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig2
-  namespace: foo
-data:
-  foo: bar
-  bar: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
