--- testdata/raw-manifests/output/files/templates/foo/configmap.2.yaml
@@ -1,12 +1,0 @@
-# Source: input/templates/foo/configmap.2.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: myconfig
-  namespace: foo
-data:
-  foo: baz
-  baz: |
-    -----BEGIN CERTIFICATE-----
-    FOO
-    -----END CERTIFICATE-----
