--- testdata/raw-manifests/output/templates/foo/configmap.2.yaml
@@ -1,1 +1,0 @@
-{{ .Files.Get "files/templates/foo/configmap.2.yaml" }}
