--- testdata/setns/input/templates/one.yaml
@@ -1,7 +1,0 @@
-# Source: templates/one.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config3
-stringData:
-  source: config3
