--- testdata/setns/input/templates/two.yaml
@@ -1,15 +1,0 @@
-# Source: templates/two.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config1
-stringData:
-  source: config1
----
-# Source: templates/two.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config2
-stringData:
-  source: config2
