--- testdata/setns/output/templates/four.yaml
@@ -1,17 +1,0 @@
-# Source: templates/four.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config5
-  namespace: void
-stringData:
-  source: config5
----
-# Source: templates/four.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config6
-  namespace: void
-stringData:
-  source: config6
