--- testdata/setns/output/templates/one.yaml
@@ -1,8 +1,0 @@
-# Source: templates/one.yaml
-apiVersion: v1
-kind: ConfigMap
-metadata:
-  name: config3
-  namespace: foo
-stringData:
-  source: config3
